Change 'unsigned long' to 'unsigned long long'
Changed long 4 bytes til long long 8 bytes in mqtt recall

+++ src/CurrentSensor/current_sensor.cpp
@@ -152,7 +152,7 @@ const unsigned long mqttRetryInterval = 5000; // Try reconnect every 5 seconds
 void reconnectMQTT() {
   // Non-blocking reconnect - only try once per call
   if (!client.connected()) {
-    unsigned long now = millis();
+    unsigned long long now = millis();
     if (now - lastMQTTAttempt >= mqttRetryInterval) {
       lastMQTTAttempt = now;
       Serial.println("Attempting MQTT connection...");
